Fix missed menu entry after browser restart
Switching permission to tun in incognito mode caused disappeared context
menu entry as well. The problem was with code recommended for Chrome
event pages, but this mode is unsupported by Firefox. Firefox extensions
are persistent (could not be temporary unloaded), so menu entry must be
created when background script is loaded.