acpContentScriptCopy: Use temporary input element when a frame is active
`document.execCommand("copy")` causes a `copy` event inside a subframe
in the case of fallback to the top page frame, so it is impossible
to overwrite `clipboardData`. Create a temporary `<textarea>` instead
despite it causes lost of active element inside the subframe.

Restore selection and active element at least in the top frame
after using a temporary input element.